Moving away database call helpers
