+++ client/src/pages/ProjectEdit/ProjectEdit.tsx
@@ -1,49 +1,10 @@
 import React, { useEffect, useState } from 'react';
 import { useParams, Link } from 'react-router-dom';
-
-const fetchProject = async (
-	{ id }: Record<string, string>,
-): Promise<Response | null> => {
-	try {
-		return fetch(`/projects/${id}`, { method: 'GET' });
-	}
-	catch (err) {
-		console.error(err);
-		return null;
-	}
-};
-
-const updateProject = async (
-	{ id, name, description }: Record<string, string>,
-): Promise<void> => {
-	try {
-		await fetch(`/projects/${id}/update`, {
-			method: 'POST',
-			headers: {
-				Accept: 'application/json',
-				'Content-Type': 'application/json',
-			},
-			body: JSON.stringify({
-				name,
-				description,
-			}),
-		});
-	}
-	catch (err) {
-		console.error(err);
-	}
-};
-
-const deleteProject = async (
-	{ id }: {id: string},
-): Promise<void> => {
-	try {
-		await fetch(`/projects/${id}/delete`, { method: 'POST' });
-	}
-	catch (err) {
-		console.error(err);
-	}
-};
+import {
+	fetchProject,
+	updateProject,
+	deleteProject,
+} from '../../functions/server-calls';
 
 // Assumes the project is done downloading.
 const ProjectEditSuccess = ({ project }: {
